Adding the progress I have made so far
Added a scanner for a civ file that reads the first line being an integer then that integer determines how many numbers there are so I created another scanner to go through the file starting at the second position until the end and store the values in an array. Then I will implement the brute force and recursive method for the function.